ai4s_agent_system/requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.1
langchain-community>=0.0.1
langchain-classic>=0.1.0
openai>=1.1.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.24.0
pytest>=7.4.0
aiohttp>=3.8.5
python-dotenv>=1.0.0
requests>=2.31.0
celery>=5.3.0
redis>=4.6.0
flower>=2.0.0 
datasets==4.3.0
tqdm==4.67.1